feat: create veritone mapper
