+++ server/src/jobs/offrePartenaire/veritone/atlas/importAtlas.ts
@@ -2,11 +2,10 @@ import { JOBPARTNERS_LABEL } from "shared/models/jobsPartners.model"
 import rawAtlasModel from "shared/models/rawAtlas.model"
 
 import config from "@/config"
-import { atlasJobToJobsPartners, ZAtlasJob } from "@/jobs/offrePartenaire/atlas/atlasMapper"
-
-import { importFromStreamInXml } from "../importFromStreamInXml"
-import { importFromUrlInXml } from "../importFromUrlInXml"
-import { rawToComputedJobsPartners } from "../rawToComputedJobsPartners"
+import { importFromStreamInXml } from "@/jobs/offrePartenaire/importFromStreamInXml"
+import { importFromUrlInXml } from "@/jobs/offrePartenaire/importFromUrlInXml"
+import { rawToComputedJobsPartners } from "@/jobs/offrePartenaire/rawToComputedJobsPartners"
+import { veritoneJobToJobsPartners, ZVeritoneJob } from "@/jobs/offrePartenaire/veritone/veritoneMapper"
 
 const rawCollectionName = rawAtlasModel.collectionName
 const offerXmlTag = "job"
@@ -23,8 +22,8 @@ export const importAtlasToComputed = async () => {
   await rawToComputedJobsPartners({
     collectionSource: rawCollectionName,
     partnerLabel: JOBPARTNERS_LABEL.ATLAS,
-    zodInput: ZAtlasJob,
-    mapper: atlasJobToJobsPartners,
+    zodInput: ZVeritoneJob,
+    mapper: (job) => veritoneJobToJobsPartners(job, JOBPARTNERS_LABEL.ATLAS),
     documentJobRoot: offerXmlTag,
   })
 }
