+++ server/src/jobs/offrePartenaire/veritone/atlas/processAtlas.ts
@@ -1,0 +1,6 @@
+import { importAtlasRaw, importAtlasToComputed } from "@/jobs/offrePartenaire/veritone/atlas/importAtlas"
+
+export const processAtlas = async () => {
+  await importAtlasRaw()
+  await importAtlasToComputed()
+}
