+++ server/src/jobs/offrePartenaire/veritone/meteojob/processMeteojob.ts
